requirements.txt
python-telegram-bot==20.7
Pillow==10.1.0
python-dotenv==1.0.0
schedule==1.2.1
